error handling

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save()
       .then((createdpost) => {
@@ -49,8 +50,12 @@ router.post(
             imagePath: createdpost.imagePath
           }
         });
+      })
+      .catch(error => {
+        res.status(500).json({
+          message: "Creating new post failed."
+        })
       });
-
   });
 
 router.get('', (req, res, next) => {
@@ -76,6 +81,11 @@ router.get('', (req, res, next) => {
         posts: fetchedPosts,
         maxPosts: count
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching posts failed."
+      })
     });
 });
 
@@ -92,14 +102,25 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post)
+    // console.log(post);
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
       .then((result) => {
         // console.log("edit post: " + result);
-        res.status(200).json({ message: 'update successful.' });
+        if (result.nModified > 0) {
+          res.status(200).json({ message: 'Update successful.' });
+        }
+        else {
+          res.status(401).json({ message: "Not Authorized." });
+        }
       })
+      .catch(error => {
+        res.status(500).json({
+          message: "Couldn't update post."
+        })
+      });
   });
 
 router.get('/:id', (req, res, next) => {
@@ -110,15 +131,30 @@ router.get('/:id', (req, res, next) => {
       } else {
         res.status(404).json({ message: 'post not found.' });
       }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching post failed."
+      })
     });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
   console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then((result) => {
       // console.log(result);
-      res.status(200).json({ message: "Post deleted." });
+      if (result.n > 0) {
+        res.status(200).json({ message: "Post deleted." });
+      }
+      else {
+        res.status(401).json({ message: "Not Authorized." });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching post failed."
+      })
     });
 });
